requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
boto3==1.35.4
botocore==1.35.4
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
deepgram-sdk==3.5.0
deprecation==2.1.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jmespath==1.0.1
marshmallow==3.22.0
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.42.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==13.0
yarl==1.9.4
